Prepare for project release to a personal Github account
Due to the small size and personal nature of this project, the
Microsoft Open Source Programs Office recommended releasing
this under a personal GitHub account instead of under a
Microsoft-managed organization. Furthermore, they recommmended
adding a note to the README explaining that this is a small
project and not formally supported by Microsoft.

This change implements the above guidance. Since this project
will not be launched under a Microsoft-managed organization and
will is not formally supported by Microsoft, this change also removes
mentions of the Microsoft Contributor License Agreement (CLA)
and Microsoft security reporting guidelines.