Data flow editor now auto-adds on selection of destination component
Having a separate add button doesn't really add
value for the data flow editor - since selecting a
destination component for a data flow is the last
bit of information we need, we can trigger the
add action on selection of the destination component to save users
the extra click.

This does make it a little harder if users accidentally
add a dataflow they didn't intend to add (since
there is no more separate form-submit event);
however, since it's only a few clicks to fix such an
error, and we assume such errors should be
uncommon, this change biases toward making intential
adds slightly faster at the expense of making
correction of unintentional adds slightly slower.